@@ -3,35 +3,17 @@ import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RequestInterceptor } from './interceptors/request.interceptor';
-import { APOLLO_OPTIONS } from 'apollo-angular';
-import { HttpLink } from 'apollo-angular/http';
-import { InMemoryCache } from '@apollo/client/core';
-
 import { throwIfAlreadyLoaded } from './guards/module-import.guard';
-import { environment } from '@env';
+import { GraphQLModule } from './graphql.module';
 
 @NgModule({
   declarations: [],
-  imports: [BrowserAnimationsModule, BrowserModule, HttpClientModule],
+  imports: [BrowserAnimationsModule, BrowserModule, GraphQLModule, HttpClientModule],
   providers: [
     {
       provide: HTTP_INTERCEPTORS,
       useClass: RequestInterceptor,
       multi: true,
-    },
-    {
-      provide: APOLLO_OPTIONS,
-      useFactory: (httpLink: HttpLink) => {
-        return {
-          cache: new InMemoryCache({
-            addTypename: false,
-          }),
-          link: httpLink.create({
-            uri: environment.pokeApiGraphQL,
-          }),
-        };
-      },
-      deps: [HttpLink],
     },
   ],
 })
