@@ -6,34 +6,31 @@ const routes: Routes = [
   {
     path: '',
     component: LayoutComponent,
-    loadChildren: () => import('app/views/home/home.module').then((m) => m.HomeModule),
+    loadChildren: () => import('./views/home/home.module').then((m) => m.HomeModule),
   },
   {
     path: 'pokemon/:id',
     component: LayoutComponent,
     loadChildren: () =>
-      import('app/views/pokemon-details/pokemon-details.module').then(
-        (m) => m.PokemonDetailsModule
-      ),
+      import('./views/pokemon-details/pokemon-details.module').then((m) => m.PokemonDetailsModule),
   },
   {
     path: 'about',
     component: LayoutComponent,
-    loadChildren: () => import('app/views/about/about.module').then((m) => m.AboutModule),
+    loadChildren: () => import('./views/about/about.module').then((m) => m.AboutModule),
   },
   {
     path: 'error',
     component: LayoutComponent,
     loadChildren: () =>
-      import('app/views/internal-server-error/internal-server-error.module').then(
+      import('./views/internal-server-error/internal-server-error.module').then(
         (m) => m.InternalServerErrorModule
       ),
   },
   {
     path: '**',
     component: LayoutComponent,
-    loadChildren: () =>
-      import('app/views/not-found/not-found.module').then((m) => m.NotFoundModule),
+    loadChildren: () => import('./views/not-found/not-found.module').then((m) => m.NotFoundModule),
   },
 ];
 
