@@ -1,6 +1,6 @@
-import Mapper from '@app/base/mapper';
-import PokemonTypeModel from '@app/services/pokemon-type/pokemon-type.model';
+import Mapper from '../../base/mapper';
 import { PokemonV2TypeEntity } from './pokemon-type.entity';
+import PokemonTypeModel from './pokemon-type.model';
 
 const PokemonTypeMapper: Mapper<PokemonV2TypeEntity, PokemonTypeModel> = {
   mapFrom: (pokemonV2TypeEntity: PokemonV2TypeEntity): PokemonTypeModel => ({
