@@ -1,18 +1,16 @@
 import { Component, OnInit } from '@angular/core';
-import { environment } from '@env';
+import { environment } from 'src/environments/environment';
 
 @Component({
   selector: 'app-about',
   templateUrl: './about.component.html',
   styleUrls: ['./about.component.scss'],
 })
-export class AboutComponent implements OnInit {
+export class AboutComponent {
   homeAngular: string = environment.homeAngular;
   homePokeApi: string = environment.homePokeApi;
   pokedexGithub: string = environment.pokedexGithub;
   keilerLinkedin: string = environment.keilerLinkedin;
 
   constructor() {}
-
-  ngOnInit(): void {}
 }
