@@ -1,8 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-
-import { SharedModule } from 'app/shared/shared.module';
-
+import { SharedModule } from 'src/app/shared/shared.module';
 import { AboutComponent } from './about.component';
 
 const routes: Routes = [
