@@ -6,13 +6,11 @@ describe('PokemonDetailsComponent', () => {
   let component: PokemonDetailsComponent;
   let fixture: ComponentFixture<PokemonDetailsComponent>;
 
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        declarations: [PokemonDetailsComponent],
-      }).compileComponents();
-    })
-  );
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [PokemonDetailsComponent],
+    }).compileComponents();
+  }));
 
   beforeEach(() => {
     fixture = TestBed.createComponent(PokemonDetailsComponent);
